fastapi-supabase/requirements.txt
fastapi[all]
uvicorn
supabase>=2.4.0
python-dotenv

python-jose[cryptography]